requirements.txt
Django==5.1.7
gunicorn==23.0.0
django-bootstrap-icons==0.9.0
django-bootstrap5==25.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-instrumentation-django
psycopg2==2.9.10
